refactor: attack speed
