--- src/components/item-display/AttackSpeed.tsx
@@ -1,20 +1,0 @@
-"use client";
-
-import React from 'react';
-import { ATTACK_SPEED_NAMES } from '@/utils/types';
-
-interface AttackSpeedProps {
-  speed: string;
-}
-
-const AttackSpeed: React.FC<AttackSpeedProps> = ({ speed }) => {
-  const displayName = ATTACK_SPEED_NAMES[speed] || speed;
-  
-  return (
-    <div className="text-gray-300 text-sm mb-2">
-      {displayName}
-    </div>
-  );
-};
-
-export default AttackSpeed;
